third_party/blink/web_tests/external/wpt/webrtc/RTCRtpSender-getStats.https-expected.txt

This is a testharness.js-based test.
[FAIL] sender.getStats() via addTransceiver should return stats report containing outbound-rtp stats
  assert_equals: Expect dictionary.selectedCandidatePairId to be string expected "string" but got "undefined"
[FAIL] sender.getStats() via addTrack should return stats report containing outbound-rtp stats
  assert_equals: Expect dictionary.selectedCandidatePairId to be string expected "string" but got "undefined"
Harness: the test ran to completion.
